Version 0.5 of the software.
This is the first main code that has working components and is where to base all other builds off of. This version includes an agreement page, navigation page, firewall configuration page, and an network test page. All 4 pages are able to connect to each other and are fully functional for their purposes.